perf: Lux and SimpleChains become package extensions
NNParameters is generic in its chain type, so the struct, the prior check and the
flat-layout assertions all stay in core; only building a chain and running its
forward pass need a backend. Those move to NoLimitsLuxExt and
NoLimitsSimpleChainsExt, which drops LoopVectorization, Polyester, SLEEFPirates,
StrideArraysCore and the Lux stack from a default install.

The backend is now selected by dispatching on the chain object via a new
_nn_model_fun, not by branching on `p.chain isa SimpleChain` inside a single
method. The old branch could not survive the split because neither extension can
name the other's types, and the comment it carried warned that dispatching on
NNParameters' chain type parameter does not reliably out-specialise the
unconstrained Lux methods when a free `Type{T}` is present. Dispatching on a
plain positional argument sidesteps that question entirely.

logit_inverse used Lux.sigmoid; it now uses StatsFuns.logistic, which was already
a dependency.

The core NNParameters constructor distinguishes the two failure modes: a wrong
chain type when a backend is loaded, and no backend at all otherwise.

using NoLimits: 6.6s -> 4.7s.
