+++ Project.toml
@@ -24,7 +24,6 @@ KernelDensity = "5ab0869b-81aa-558d-bb23-cbf5423bbe9b"
 LineSearches = "d3d80556-e9d4-5f37-9878-2ab0fcc64255"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
-Lux = "b2108857-7c20-44ae-9111-449ecde12c47"
 MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d"
 Optimisers = "3bd65402-5787-11e9-1adc-39752487f4e2"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
@@ -36,7 +35,6 @@ Roots = "f2b01f46-fcfa-551c-844a-d8ac1e96c665"
 RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 SciMLStructures = "53ae85a6-f571-4167-b2af-e1d143709226"
-SimpleChains = "de6bee2f-e2f4-4ec7-b6ed-219cc6f6e9e5"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
@@ -50,9 +48,11 @@ Copulas = "ae264745-0b69-425e-9d9d-cf662c5eec93"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 Latexify = "23fbe1c1-3f47-55db-b15f-69d7ec21a316"
 LikelihoodProfiler = "93acb638-a083-5915-8dce-d129bc6a3f59"
+Lux = "b2108857-7c20-44ae-9111-449ecde12c47"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 OptimizationNLopt = "4e6fcdb7-1186-4e1f-a706-475e75c168bb"
 SciMLSensitivity = "1ed8b502-d754-442c-8d5d-10ac956f44a1"
+SimpleChains = "de6bee2f-e2f4-4ec7-b6ed-219cc6f6e9e5"
 
 [extensions]
 NoLimitsCSVExt = "CSV"
@@ -60,8 +60,10 @@ NoLimitsCopulasExt = "Copulas"
 NoLimitsJLD2Ext = "JLD2"
 NoLimitsLatexifyExt = "Latexify"
 NoLimitsLikelihoodProfilerExt = ["LikelihoodProfiler", "OptimizationNLopt"]
+NoLimitsLuxExt = "Lux"
 NoLimitsMakieExt = "Makie"
 NoLimitsSciMLSensitivityExt = "SciMLSensitivity"
+NoLimitsSimpleChainsExt = "SimpleChains"
 
 [compat]
 AdvancedVI = "0.6.2, 0.7"
